- Fixed .antialt disable confirmation embed color - Fixed weighted warning punishment

+++ src/NadekoBot/Db/Models/Warning.cs
@@ -9,5 +9,5 @@ public class Warning : DbEntity
     public bool Forgiven { get; set; }
     public string ForgivenBy { get; set; }
     public string Moderator { get; set; }
-    public int Weight { get; set; }
+    public long Weight { get; set; }
 }
